Add Fish creature prefab and update DEV scene
Added a new Fish prefab with associated meta files and integrated it into the _DEV_Yanis scene. Updated the scene to include a floor with NavMeshSurface and placed the Fish prefab in the environment. Also made minor adjustments to creature scripts and NavMesh/quality settings to support the new content.